Reorganizes Parser into specialized parsers.
Changes the links API.

Ref #98
