+++ lib/meta_inspector.rb
@@ -4,6 +4,11 @@ require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/excep
 require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/request'))
 require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/url'))
 require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/parser'))
+require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/parsers/base'))
+require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/parsers/images'))
+require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/parsers/links'))
+require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/parsers/meta_tags'))
+require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/parsers/texts'))
 require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/document'))
 require File.expand_path(File.join(File.dirname(__FILE__), 'meta_inspector/version'))
