+++ lib/meta_inspector/version.rb
@@ -1,3 +1,3 @@
 module MetaInspector
-  VERSION = "3.3.0"
+  VERSION = "4.0.0.rc1"
 end
